Move toExpr out of Declaration
